requirements.txt
gradio
psycopg2-binary
pgvector
dashscope
numpy
python-dotenv
oss2
redis
openai
aiohttp
